Queuing messages received before `_registerListeners`
It is possible for RTCDataChannels to receive messages after the
`datachannel` event, but before `_registerListeners` is called. In
this scenerio, those messages are dropped since no message listener
has been set.

To mitigate this, a temporary message listener is set during the
`datachannel` event handling that pushes messages onto a queue
when they are received. When `_registerListeners` is called,
the temporary listener is replaced with the original and the
queue is processed.